requirements.txt
altgraph==0.17.2
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
future==0.18.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
macholib==1.15.2
MarkupSafe==2.1.1
pathlib==1.0.1
pipenv==2022.12.19
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.2
six==1.15.0
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
Werkzeug==2.2.2
zipp==3.11.0
